backend: split null to a file
