+++ pkg/backend/null_test.go
@@ -1,0 +1,33 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright (c) 2022 Dell Inc, or its subsidiaries.
+
+// Package backend implememnts the BackEnd APIs (network facing) of the storage Server
+package backend
+
+import (
+	"testing"
+)
+
+func TestBackEnd_CreateNullDebug(t *testing.T) {
+
+}
+
+func TestBackEnd_DeleteNullDebug(t *testing.T) {
+
+}
+
+func TestBackEnd_UpdateNullDebug(t *testing.T) {
+
+}
+
+func TestBackEnd_ListNullDebugs(t *testing.T) {
+
+}
+
+func TestBackEnd_GetNullDebug(t *testing.T) {
+
+}
+
+func TestBackEnd_NullDebugStats(t *testing.T) {
+
+}
